+++ tests/assets/testpackage/three/beta.py
@@ -1,0 +1,5 @@
+from testpackage.one import alpha
+
+
+def foo():
+    return alpha.BAR